# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Best Courses Ever API
  version: 1.0.0
  description: REST API для образовательной платформы Best Courses Ever
servers:
  - url: http://localhost:3008
    description: Локальный сервер разработки
paths:
  /api/v1/users:
    get:
      summary: Получить список всех пользователей
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '500':
          description: Ошибка сервера
    post:
      summary: Создать нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Некорректные данные
        '500':
          description: Ошибка сервера
  /api/v1/users/{id}:
    get:
      summary: Получить пользователя по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Пользователь найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Пользователь не найден
        '500':
          description: Ошибка сервера
    put:
      summary: Обновить пользователя по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Пользователь обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Пользователь не найден
        '500':
          description: Ошибка сервера
    delete:
      summary: Удалить пользователя по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Пользователь удалён
        '404':
          description: Пользователь не найден
        '500':
          description: Ошибка сервера

  /api/v1/courses:
    get:
      summary: Получить список всех курсов
      responses:
        '200':
          description: Список курсов
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Course'
        '500':
          description: Ошибка сервера
    post:
      summary: Создать новый курс
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreate'
      responses:
        '201':
          description: Курс создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          description: Некорректные данные
        '500':
          description: Ошибка сервера

  /api/v1/courses/{id}:
    get:
      summary: Получить курс по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Курс найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '404':
          description: Курс не найден
        '500':
          description: Ошибка сервера
    put:
      summary: Обновить курс по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseUpdate'
      responses:
        '200':
          description: Курс обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '404':
          description: Курс не найден
        '500':
          description: Ошибка сервера
    delete:
      summary: Удалить курс по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Курс удалён
        '404':
          description: Курс не найден
        '500':
          description: Ошибка сервера

components:
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: ['user','author','admin']
        createdAt:
          type: string
        updatedAt:
          type: string
    UserCreate:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: ['user','author','admin']
    UserUpdate:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: ['user','author','admin']
    Course:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
        tags:
          type: array
          items:
            type: string
        author_id:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
        updatedAt:
          type: string
    CourseCreate:
      type: object
      required:
        - title
        - description
        - difficulty
        - tags
        - author_id
      properties:
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
        tags:
          type: array
          items:
            type: string
        author_id:
          type: string
    CourseUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
        tags:
          type: array
          items:
            type: string
        author_id:
          type: string